Renamed the facebook adapter to audienceNetwork.
