+++ adapters/audienceNetwork/facebook_test.go
@@ -1,4 +1,4 @@
-package facebook
+package audienceNetwork
 
 import (
 	"bytes"
