+++ adapters/audienceNetwork/info.yaml
