+++ static/bidder-params/audienceNetwork.json
@@ -1,7 +1,7 @@
 {
   "$schema": "http://json-schema.org/draft-04/schema#",
-  "title": "Facebook Adapter Params",
-  "description": "A schema which validates params accepted by the Facebook adapter",
+  "title": "Facebook Audience Network Adapter Params",
+  "description": "A schema which validates params accepted by the Facebook Audience Network adapter",
   "type": "object",
   "properties": {
     "placementId": {
